agri app phase 2

+++ lib/weather/models/weather.dart
@@ -144,7 +144,7 @@ class Wind {
   });
 
   factory Wind.fromJson(Map<String, dynamic> json) => Wind(
-        speed: json['speed'] ?? 0.0,
+        speed: (json['speed']).toDouble() ?? 0.0,
         deg: json['deg'] ?? 0,
         gust: json['gust'],
       );
